requirements.txt
python-dotenv==1.2.1
google-generativeai==0.8.6
google-auth==2.48.0
google-api-python-client==2.190.0
gspread==6.2.1
tenacity==9.1.4